+++ packages/terra-abstract-modal/tests/wdio/abstract-modal-spec.js
@@ -4,16 +4,18 @@ describe('Abstract Modal', () => {
   describe('Append Class', () => {
     before(() => browser.url('/#/raw/tests/terra-abstract-modal/abstract-modal/abstract-modal-append-class'));
 
-    Terra.should.matchScreenshot({ selector: 'div[role="dialog"]' });
-    Terra.should.beAccessible();
+    Terra.should.validateElement({ selector: 'div[role="dialog"]' });
   });
 
   describe('Content Overflow', () => {
-    before(() => browser.url('/#/raw/tests/terra-abstract-modal/abstract-modal/abstract-modal-content-overflow'));
-    const overflowModal = { height: 1684, width: 1870 };
+    before(() => {
+      browser.url('/#/raw/tests/terra-abstract-modal/abstract-modal/abstract-modal-content-overflow');
+      browser.setViewportSize({ height: 1684, width: 1870 });
+    });
 
-    Terra.should.matchScreenshot({ viewports: [overflowModal], selector: 'div[role="dialog"]' });
-    Terra.should.beAccessible({ viewports: [overflowModal] });
+    after(() => browser.setViewportSize(Terra.viewports('medium')[0]));
+
+    Terra.should.validateElement({ selector: 'div[role="dialog"]' });
     Terra.should.themeCombinationOfCustomProperties({
       testName: 'themed',
       properties: {
@@ -21,7 +23,6 @@ describe('Abstract Modal', () => {
         '--terra-abstract-modal-horizontal-inset': '50px',
       },
       selector: 'div[role="dialog"]',
-      viewports: [overflowModal],
     });
   });
 
@@ -35,14 +36,13 @@ describe('Abstract Modal', () => {
   describe('Disable Close On Esc', () => {
     before(() => browser.url('/#/raw/tests/terra-abstract-modal/abstract-modal/abstract-modal-disable-close-on-esc'));
 
-    Terra.should.matchScreenshot('modal present', { selector: 'div[role="dialog"]' });
+    Terra.should.validateElement('modal present', { selector: 'div[role="dialog"]' });
 
     it('attempts to close modal with ESC', () => {
       browser.keys('Escape');
     });
 
-    Terra.should.matchScreenshot('modal present', { selector: 'div[role="dialog"]' });
-    Terra.should.beAccessible();
+    Terra.should.validateElement('modal present', { selector: 'div[role="dialog"]' });
   });
 
   describe('Within Modal Focus Handling', () => {
@@ -52,27 +52,28 @@ describe('Abstract Modal', () => {
       expect(browser.hasFocus('[aria-modal="true"][role="dialog"]')).to.be.equal(true);
     });
 
-    Terra.should.matchScreenshot('modal is focused on open', { selector: '#root' });
+    Terra.should.validateElement('modal is focused on open', { selector: '#root' });
 
     it('focuses on interactive elements within the modal', () => {
       browser.keys(['Tab']);
       expect(browser.hasFocus('#focus-button')).to.be.equal(true);
     });
 
-    Terra.should.matchScreenshot('modal button focused', { selector: '#root' });
+    Terra.should.validateElement('modal button focused', { selector: '#root' });
 
     it('does not focus on interactive content within the app when the modal is open - tab cycle forward', () => {
       browser.keys(['Tab']);
+      expect(browser.hasFocus('#focus-button')).to.be.equal(false);
     });
 
-    Terra.should.matchScreenshot('focused shifted outside the end of the modal', { selector: '#root' });
+    Terra.should.validateElement('focused shifted outside the end of the modal', { selector: '#root' });
 
     it('shifts focus back onto interactive elements within the modal', () => {
       browser.keys(['Shift', 'Tab']);
       expect(browser.hasFocus('#focus-button')).to.be.equal(true);
     });
 
-    Terra.should.matchScreenshot('modal button focused again', { selector: '#root' });
+    Terra.should.validateElement('modal button focused again', { selector: '#root' });
   });
 
   describe('Outside Modal Focus Handling', () => {
@@ -80,12 +81,14 @@ describe('Abstract Modal', () => {
 
     it('does not focus on interactive content within the app when modal is open and focus is shifted backwards from the modal dialog', () => {
       browser.keys(['Shift', 'Shift', 'Tab']);
+      expect(browser.hasFocus('#focus-button')).to.be.equal(false);
     });
 
     Terra.should.matchScreenshot('focused shifted outside the beginning of the modal', { selector: '#root' });
 
     it('does not focus on interactive content within the app when modal is open and focus is shifted forwards from the modal dialog', () => {
       browser.keys(['Tab', 'Tab']);
+      expect(browser.hasFocus('#focus-button')).to.be.equal(false);
     });
 
     Terra.should.matchScreenshot('focused shifted outside the end of the modal', { selector: '#root' });
@@ -94,65 +97,61 @@ describe('Abstract Modal', () => {
   describe('Disable Close On Outside Click', () => {
     before(() => browser.url('/#/raw/tests/terra-abstract-modal/abstract-modal/abstract-modal-disable-close-on-outside-click'));
 
-    Terra.should.matchScreenshot('modal present', { selector: 'div[role="dialog"]' });
+    Terra.should.validateElement('modal present', { selector: 'div[role="dialog"]' });
 
     it('attempts to close modal with outside click', () => {
       browser.leftClick('button');
     });
-    Terra.should.matchScreenshot('modal present', { selector: 'div[role="dialog"]' });
-    Terra.should.beAccessible();
+    Terra.should.validateElement('modal present', { selector: 'div[role="dialog"]' });
   });
 
   describe('Is Fullscreen', () => {
     before(() => browser.url('/#/raw/tests/terra-abstract-modal/abstract-modal/abstract-modal-is-fullscreen'));
 
-    Terra.should.matchScreenshot('modal present', { selector: 'div[role="dialog"]' });
+    Terra.should.validateElement('modal present', { selector: 'div[role="dialog"]' });
+
     it('clicks outside to close modal', () => {
       browser.moveToObject('[class*="ModalOverlay"]', 235, 5)
         .leftClick();
     });
-    Terra.should.matchScreenshot('modal closed');
+    Terra.should.validateElement('modal closed');
 
     it('opens the modal', () => {
       browser.click('button');
     });
-    Terra.should.matchScreenshot('modal present', { selector: 'div[role="dialog"]' });
+    Terra.should.validateElement('modal present', { selector: 'div[role="dialog"]' });
 
     it('closes modal on ESC', () => {
       browser.keys('Escape');
     });
-    Terra.should.matchScreenshot('modal closed, with focus');
-    Terra.should.beAccessible();
+    Terra.should.validateElement('modal closed, with focus');
   });
 
   describe('Is Open', () => {
     before(() => browser.url('/#/raw/tests/terra-abstract-modal/abstract-modal/abstract-modal-is-open'));
 
-    Terra.should.matchScreenshot('modal closed');
+    Terra.should.validateElement('modal closed');
+
     it('clicks to open modal', () => {
       browser.click('button');
     });
-    Terra.should.matchScreenshot({ selector: 'div[role="dialog"]' });
+    Terra.should.validateElement({ selector: 'div[role="dialog"]' });
 
     it('closes the modal on ESC', () => {
       browser.keys('Escape');
     });
-    Terra.should.matchScreenshot('modal closed, with focus');
-
-    Terra.should.beAccessible();
+    Terra.should.validateElement('modal closed, with focus');
   });
 
   describe('No Focusable Content', () => {
     before(() => browser.url('/#/raw/tests/terra-abstract-modal/abstract-modal/abstract-modal-no-focusable-content'));
 
-    Terra.should.matchScreenshot({ selector: 'div[role="dialog"]' });
-    Terra.should.beAccessible();
+    Terra.should.validateElement({ selector: 'div[role="dialog"]' });
   });
 
   describe('Override Role', () => {
     before(() => browser.url('/#/raw/tests/terra-abstract-modal/abstract-modal/abstract-modal-override-role'));
 
-    Terra.should.matchScreenshot({ selector: 'div[role="alertdialog"]' });
-    Terra.should.beAccessible();
+    Terra.should.validateElement({ selector: 'div[role="alertdialog"]' });
   });
 });