+++ packages/terra-abstract-modal/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Update tests to use `Terra.should.validateElement()`
 
 3.0.0 - (May 13, 2019)
 ------------------