+++ packages/terra-application-utility/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Update tests to use `Terra.should.validateElement()`
 
 2.10.0 - (March 15, 2019)
 ------------------